+++ src/server/CMakeLists.txt
@@ -392,8 +392,8 @@ set(SRCS
 	src/ipc-command-handler.hpp
 	src/ipc-command-handler.cpp
 
-	src/interval-scheduler.hpp
-	src/interval-scheduler.cpp
+	src/extension-interval-scheduler.hpp
+	src/extension-interval-scheduler.cpp
 
 	src/log/message-handler.cpp
