+++ src/server/src/services/extension-registry/extension-manifest.cpp
@@ -162,37 +162,30 @@ CommandArgument ExtensionManifest::parseArgumentFromObject(const QJsonObject &ob
   return arg;
 }
 
-static std::optional<std::chrono::seconds> parseInterval(const std::string &str) {
-  if (str.empty()) return std::nullopt;
+static std::expected<std::chrono::seconds, QString> parseInterval(const std::string &str) {
+  if (str.empty()) return std::unexpected(QString("interval is empty"));
 
   char unit = str.back();
 
   std::int64_t value = 0;
   auto [ptr, ec] = std::from_chars(str.data(), str.data() + str.size() - 1, value);
-  if (ec != std::errc{} || ptr != str.data() + str.size() - 1) return std::nullopt;
+  if (ec != std::errc{} || ptr != str.data() + str.size() - 1)
+    return std::unexpected(QString("invalid interval format: %1").arg(str.c_str()));
 
   using namespace std::chrono;
 
   seconds secs;
 
   switch (unit) {
-  case 's':
-    secs = seconds(value);
-    break;
-  case 'm':
-    secs = duration_cast<seconds>(minutes(value));
-    break;
-  case 'h':
-    secs = duration_cast<seconds>(hours(value));
-    break;
-  case 'd':
-    secs = duration_cast<seconds>(hours(value * 24));
-    break;
-  default:
-    return std::nullopt;
+  case 's': secs = seconds(value); break;
+  case 'm': secs = duration_cast<seconds>(minutes(value)); break;
+  case 'h': secs = duration_cast<seconds>(hours(value)); break;
+  case 'd': secs = duration_cast<seconds>(hours(value * 24)); break;
+  default: return std::unexpected(QString("unknown interval unit: %1").arg(unit));
   }
 
-  if (secs < seconds(5)) return std::nullopt;
+  if (secs < seconds(5))
+    return std::unexpected(QString("interval must be at least 5s, got %1s").arg(secs.count()));
 
   return secs;
 }
@@ -217,7 +210,12 @@ ExtensionManifest::Command ExtensionManifest::parseCommandFromObject(const QJson
   }
 
   if (obj.contains("interval") && command.mode == CommandModeNoView) {
-    command.interval = parseInterval(obj.value("interval").toString().toStdString());
+    auto interval = parseInterval(obj.value("interval").toString().toStdString());
+    if (interval) {
+      command.interval = *interval;
+    } else {
+      qWarning() << "Failed to parse interval for command" << command.name << "-" << interval.error();
+    }
   }
 
   for (const auto &obj : obj.value("preferences").toArray()) {
